requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.14
docstring-parser==0.15
fire==0.5.0
frozenlist==1.3.3
fsspec==2023.6.0
gcsfs==2023.6.0
google-api-core==2.10.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-aiplatform==1.26.1
google-cloud-bigquery==3.11.2
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-notebooks==1.4.4
google-cloud-pipeline-components==1.0.8
google-cloud-resource-manager==1.6.3
google-cloud-storage==1.44.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.48.2
httplib2==0.22.0
idna==3.4
joblib==1.2.0
jsonschema==3.2.0
kfp==1.8.12
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kubernetes==18.20.0
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.10.1
Shapely==1.8.5.post1
six==1.16.0
strip-hints==0.1.10
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.16
websocket-client==1.6.1
wrapt==1.15.0
yarl==1.9.2
